requirements.txt
python-dotenv
streamlit
pypdf
apify-client
pyttsx3
SpeechRecognition
pyaudio

langchain
langchain-openai
langchain-nvidia-ai-endpoints
langchain-core
langchain-community
langgraph
langgraph-supervisor
langchain-google-community
google_auth_oauthlib

langchain-tavily
langchain-pinecone

plotly
sqlalchemy
psycopg2-binary

pandas
numpy
matplotlib